public_dropin_notebook_environments/python311_notebook/requirements.txt
altair==4.2.0
cloudpickle==2.2.1
dask==2022.9.2
datarobot-bp-workshop==0.2.6
datarobot-drum==1.10.19
datarobot-mlops-connected-client<10.0.0
datarobot-model-metrics==0.5.4
datarobot-predict==1.5.1
datarobot==3.4.0
dummyPy==0.3
eli5==0.13.0
gensim==4.3.2
graphviz==0.20.1
imblearn
joblib==1.2.0
latexify-py==0.4.2
lime==0.2.0.1
matplotlib==3.6.1
mistune==2.0.4
numpy==1.26.2
opencv-python-headless==4.8.1.78
openpyxl==3.0.10
pandas==1.5.1
plotly==5.18.0
pydantic==2.5.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.0
shap==0.41.0
snowflake-connector-python[secure-local-storage,pandas]==3.7.1
snowflake-snowpark-python==1.11.1
spacy==3.7.2
statsmodels==0.14.1
tqdm==4.64.1
umap-learn==0.5.3
wordcloud==1.9.3
xgboost==1.6.2
yellowbrick==1.5
boto3==1.34.90
sagemaker==2.216.1
databricks-connect==13.3.3